Elie/requirements.txt
Pillow
pyperclip
opencv-python
pyscreenshot
pycryptodome
pyxhook
tendo

